Make UpdatePackageRunner compatible with 7.5
The UpdateHelper no longer exposes the SaveInstallationMessage (see #30)
in 7.5. Suggested fix that is backwards compatible with previous version
is to implement a local equivalent of that method.

Also added diagnostic output when saving of the installation messages fails.
I have seen it a few times due to an exception during the installation
attempt. If stack trace includes lines from the obfuscated nexus assembly
the XML may end up having special characters that would fail serialization
and fail the entire SaveInstallationMessages command.